requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly-express